Move tests in their own file and remove macros
