+++ parity-crypto/src/pbkdf2/mod.rs
@@ -29,25 +29,4 @@ pub fn sha512(iter: u32, salt: Salt, sec: Secret, out: &mut [u8; 64]) {
 
 
 #[cfg(test)]
-::tests_pbkdf2!();
-
-#[cfg(test)]
-#[macro_export]
-macro_rules! tests_pbkdf2 {
-	() => {
-
-mod test {
-	use super::*;
-	#[test]
-	fn basic_test() {
-		let mut dest = [0;32];
-		let salt = [5;32];
-		let secret = [7;32];
-		sha256(3, Salt(&salt[..]), Secret(&secret[..]), &mut dest);
-		let res = [242, 33, 31, 124, 36, 223, 179, 185, 206, 175, 190, 253, 85, 33, 23, 126, 141, 29, 23, 97, 66, 63, 51, 196, 27, 255, 135, 206, 74, 137, 172, 87];
-		assert_eq!(res, dest);
-	}
-}
-
-}
-}
+mod test;
